+++ packages/vscode-extension/src/registrationService.ts
