--- packages/vscode-extension/src/test.ts
@@ -1,1 +1,0 @@
-document.addEventListener('test', () => {});
